+++ MapboxCoreNavigation/BorderCrossing.swift
@@ -1,0 +1,23 @@
+
+import Foundation
+import MapboxNavigationNative
+import MapboxDirections
+
+extension AdministrativeRegion {
+    init(_ adminInfo: RouteAlertAdminInfo) {
+        self.init(countryCode: adminInfo.iso_3166_1, countryCodeAlpha3: adminInfo.iso_3166_1_alpha3)
+    }
+}
+
+/**
+ `BorderCrossingInfo` encapsulates a border crossing, specifying crossing region codes.
+ */
+public struct BorderCrossing {
+    public let from: AdministrativeRegion
+    public let to: AdministrativeRegion
+    
+    init(_ borderCrossing: RouteAlertBorderCrossingInfo) {
+        from = AdministrativeRegion(borderCrossing.from)
+        to = AdministrativeRegion(borderCrossing.to)
+    }
+}